BirthdayCalendar converting date service to be a regular class and creating an abstract class for it
